refactor: compile schematics into a dist folder
