+++ package.json
@@ -156,7 +156,7 @@
     "eslint-plugin-unicorn": "41.0.1",
     "eslint-plugin-unused-imports": "^2.0.0",
     "husky": "^7.0.4",
-    "intershop-schematics": "file:schematics/src",
+    "intershop-schematics": "file:schematics",
     "jest": "^27.5.1",
     "jest-extended": "^2.0.0",
     "jest-marbles": "^3.0.2",
