+++ schematics/.gitignore
@@ -1,7 +1,5 @@
 # outputs
-src/**/*.js
-!src/utils/lint-fix-mock.js
-src/**/*.d.ts
+dist
 
 # dependencies
 package-lock.json
