+++ schematics/src/helpers/add-destroy/factory.ts
@@ -1,11 +1,10 @@
 import { Rule, SchematicsException, chain } from '@angular-devkit/schematics';
 import { buildDefaultPath, getWorkspace } from '@schematics/angular/utility/workspace';
+import { PWAAddDestroySubjectToComponentOptionsSchema as Options } from 'schemas/helpers/add-destroy/schema';
 import { Scope } from 'ts-morph';
 
 import { applyLintFix } from '../../utils/lint-fix';
 import { createTsMorphProject } from '../../utils/ts-morph';
-
-import { PWAAddDestroySubjectToComponentOptionsSchema as Options } from './schema';
 
 export function add(options: Options): Rule {
   return async host => {
