+++ schematics/src/helpers/customized-copy/factory.ts
@@ -3,6 +3,7 @@ import { Rule, SchematicsException, chain } from '@angular-devkit/schematics';
 import { tsquery } from '@phenomnomnominal/tsquery';
 import { getWorkspace } from '@schematics/angular/utility/workspace';
 import { basename, join } from 'path';
+import { CustomizedCopyOptionsSchema as Options } from 'schemas/helpers/customized-copy/schema';
 import * as ts from 'typescript';
 
 import { applyNameAndPath, determineArtifactName, findDeclaringModule } from '../../utils/common';
@@ -10,8 +11,6 @@ import { readIntoSourceFile } from '../../utils/filesystem';
 import { applyLintFix } from '../../utils/lint-fix';
 import { addDeclarationToNgModule } from '../../utils/registration';
 import { updateComponentClassName, updateComponentDecorator, updateComponentSelector } from '../move-component/factory';
-
-import { CustomizedCopyOptionsSchema as Options } from './schema';
 
 export function customize(options: Options): Rule {
   return async host => {
