--- schematics/src/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "intershop-schematics",
-  "version": "0.0.1",
-  "private": true,
-  "description": "Angular CLI schematics for Intershop PWA",
-  "keywords": ["schematics", "intershop"],
-  "schematics": "./collection.json",
-  "peerDependencies": {
-    "@angular-devkit/core": ">=9.0.0",
-    "@angular-devkit/schematics": ">=9.0.0",
-    "@schematics/angular": ">=9.0.0",
-    "@phenomnomnominal/tsquery": ">=4.0.0",
-    "@angular/cli": ">=9.0.0"
-  }
-}
