+++ schematics/src/page/factory.ts
@@ -12,6 +12,7 @@ import {
   url,
 } from '@angular-devkit/schematics';
 import { buildDefaultPath, getWorkspace } from '@schematics/angular/utility/workspace';
+import { PWAPageOptionsSchema as Options } from 'schemas/page/schema';
 import { forEachToken, getChildOfKind } from 'tsutils';
 import * as ts from 'typescript';
 
@@ -19,8 +20,6 @@ import { applyNameAndPath, detectExtension, determineArtifactName } from '../uti
 import { readIntoSourceFile } from '../utils/filesystem';
 import { applyLintFix } from '../utils/lint-fix';
 import { addImportToFile } from '../utils/registration';
-
-import { PWAPageOptionsSchema as Options } from './schema';
 
 function addRouteToArray(
   options: { name?: string; routingModule?: string; child?: string; lazy?: boolean },
