+++ schematics/src/page/factory_spec.ts
@@ -1,9 +1,8 @@
 import { UnitTestTree } from '@angular-devkit/schematics/testing';
 import { lastValueFrom } from 'rxjs';
 
+import { PWAPageOptionsSchema as Options } from '../../dist/page/schema';
 import { createApplication, createSchematicRunner } from '../utils/testHelper';
-
-import { PWAPageOptionsSchema as Options } from './schema';
 
 describe('Page Schematic', () => {
   const schematicRunner = createSchematicRunner();
