+++ schematics/src/store/factory.ts
@@ -11,6 +11,7 @@ import {
   url,
 } from '@angular-devkit/schematics';
 import { getWorkspace } from '@schematics/angular/utility/workspace';
+import { PWAStoreOptionsSchema as Options } from 'schemas/store/schema';
 import { forEachToken } from 'tsutils';
 import * as ts from 'typescript';
 
@@ -18,8 +19,6 @@ import { applyNameAndPath, determineArtifactName } from '../utils/common';
 import { readIntoSourceFile } from '../utils/filesystem';
 import { applyLintFix } from '../utils/lint-fix';
 import { insertImport } from '../utils/registration';
-
-import { PWAStoreOptionsSchema as Options } from './schema';
 
 // eslint-disable-next-line complexity
 export async function determineStoreLocation(
